backend/requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-cloud-vision>=3.4.0
reportlab>=4.0.0
openai>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
